Sanity check the existence of help files.
The sanity checking is a hack.  Overriding manpages.include? could test
for the existence of the pre-built file, and fallback to groff.

Despite troff'd output included in the gem, preference is given to
invoking groff every time (instead of using that pre-built output).

(This could possibly be improved if the MANPATH PR is merged.)